chore/#22: 모듈분리 및 워크스페이스 추가
