+++ MLS/DomainInterface/DomainInterface/Entity/NetworkLayer/NetworkEntity.swift
