+++ MLS/DomainInterface/DomainInterface/Entity/NetworkLayer/ResponsableEndPoint.swift
