+++ MLS/MLS.xcworkspace/contents.xcworkspacedata
@@ -1,4 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Workspace
    version = "1.0">
+   <FileRef
+      location = "group:MLS.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "container:Presentation/Presentation.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:DomainInterface/DomainInterface.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Domain/Domain.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Data/Data.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Core/Core.xcodeproj">
+   </FileRef>
 </Workspace>
