--- MLS/MLSData/MLSData/MLSData.h
@@ -1,18 +1,0 @@
-//
-//  MLSData.h
-//  MLSData
-//
-//  Created by SeoJunYoung on 4/14/25.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for MLSData.
-FOUNDATION_EXPORT double MLSDataVersionNumber;
-
-//! Project version string for MLSData.
-FOUNDATION_EXPORT const unsigned char MLSDataVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <MLSData/PublicHeader.h>
-
-
