--- MLS/MLSDomain/MLSDomain/MLSDomain.h
@@ -1,18 +1,0 @@
-//
-//  MLSDomain.h
-//  MLSDomain
-//
-//  Created by SeoJunYoung on 4/14/25.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for MLSDomain.
-FOUNDATION_EXPORT double MLSDomainVersionNumber;
-
-//! Project version string for MLSDomain.
-FOUNDATION_EXPORT const unsigned char MLSDomainVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <MLSDomain/PublicHeader.h>
-
-
