--- MLS/MLSPresentation/MLSPresentation/MLSPresentation.h
@@ -1,18 +1,0 @@
-//
-//  MLSPresentation.h
-//  MLSPresentation
-//
-//  Created by SeoJunYoung on 4/14/25.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for MLSPresentation.
-FOUNDATION_EXPORT double MLSPresentationVersionNumber;
-
-//! Project version string for MLSPresentation.
-FOUNDATION_EXPORT const unsigned char MLSPresentationVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <MLSPresentation/PublicHeader.h>
-
-
